ZooAuthenticator Project
This project was originally based in Java, it has been migrated to the Python language and enhanced by placing the .csv data into a Pymongo Database with password protection. The application allows displaying all data, sorted data, creation and deletion of records, and the ability to search through the records by name.